PyData Helsinki banner and logo (#2)
PyData Helsinki logo file, which is a basically the PyData logo from https://pydata.org/start-a-meetup/, vectorized, and added "Helsinki" text underneath it.

Banner GIMP file, which uses a photo by Tapio Haaja from Unsplash as a background, and adds the PyData Helsinki logo on top of it.

Png versions of the PyData Helsinki banner (exported from the GIMP file) in few sizes:

1920 x 1080 (16:9)
1440 x 1080 (4:3)
1080 x 1080 (1:1)